Create 21 update
#Description
Prompt the user for a file name and open file entered.
#Variables
String is used to name what will be used to open desired file.
#In/Out
The input by the user is for a file name. Output is to read the file.
output of file.
#Decision
Use of if statement to notify user that file has been opened.
#Iteration
writing in a for loop to read data until file end.
#File I/O
Using fstream to read file type for color file code will use.